FIX README, 예외처리

+++ src/answer/answer.service.ts
@@ -30,7 +30,10 @@ export class AnswerService {
     const question = await this.questionRepository.findOne({
       where: { id: questionId },
     });
-    if (question === null) throw new ApolloError('문항이 존재하지 않습니다.');
+    const materServey = await this.serveyRepository.findOne({
+      where: { id: question.serveyId },
+    });
+    if (question === null) throw new ApolloError('존재하지 않는 문항입니다.');
     if (question.isObjective === false)
       throw new ApolloError('해당 문항은 객관식이 아닙니다.');
     const answers = await this.answerRepository.find({
@@ -38,6 +41,10 @@ export class AnswerService {
     });
     if (answers.length === 10)
       throw new ApolloError('한 문항에 답변은 10개 이하로 해야합니다.');
+    if (materServey.isUsed === true)
+      throw new ApolloError(
+        '이미 한번 이상 응답된 설문지 입니다. 수정 및 삭제할 수 없습니다.',
+      );
 
     const listNumber = String(answers.length + 1);
     const answer = new Answer();
@@ -66,7 +73,7 @@ export class AnswerService {
     });
     if (materServey.isUsed === true)
       throw new ApolloError(
-        '한 번 이상 진행된 설문이라 내용을 변경할 수 없습니다.',
+        '이미 한번 이상 응답된 설문지 입니다. 수정 및 삭제할 수 없습니다.',
       );
     //
     const updateing = await this.answerRepository.update(answer, toUpdate);
@@ -74,6 +81,19 @@ export class AnswerService {
   }
 
   async delete(id: number): Promise<true> {
+    const answer = await this.answerRepository.findOne({ where: { id } });
+    if (!answer) throw new ApolloError('존재하지 않는 답변입니다.');
+
+    const momQuestion = await this.questionRepository.findOne({
+      where: { id: answer.questionId },
+    });
+    const materServey = await this.serveyRepository.findOne({
+      where: { id: momQuestion.serveyId },
+    });
+    if (materServey.isUsed === true)
+      throw new ApolloError(
+        '이미 한번 이상 응답된 설문지 입니다. 수정 및 삭제할 수 없습니다.',
+      );
     await this.answerRepository.delete({ id });
     return true;
   }
